+++ lib/flutter_login.dart
@@ -6,6 +6,7 @@ import 'package:flutter/scheduler.dart' show timeDilation;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_login/src/models/login_user_type.dart';
 import 'package:provider/provider.dart';
 import 'src/providers/login_theme.dart';
 import 'src/widgets/null_widget.dart';
@@ -20,22 +21,22 @@ import 'src/widgets/fade_in.dart';
 import 'src/widgets/hero_text.dart';
 import 'src/widgets/gradient_box.dart';
 export 'src/models/login_data.dart';
+export 'src/models/login_user_type.dart';
 export 'src/providers/login_messages.dart';
 export 'src/providers/login_theme.dart';
 import 'src/constants.dart';
 
 class LoginProvider {
   final IconData icon;
+  final String label;
   final ProviderAuthCallback callback;
 
-  LoginProvider({@required this.icon, @required this.callback})
-      : assert((icon != null && callback != null),
-            ' callback and icon should not be null');
+  LoginProvider({required this.icon, required this.callback, this.label = ''});
 }
 
 class _AnimationTimeDilationDropdown extends StatelessWidget {
   _AnimationTimeDilationDropdown({
-    @required this.onChanged,
+    required this.onChanged,
     this.initialValue = 1.0,
   });
 
@@ -65,7 +66,7 @@ class _AnimationTimeDilationDropdown extends StatelessWidget {
               ),
               itemExtent: 30.0,
               backgroundColor: Colors.white,
-              onSelectedItemChanged: onChanged,
+              onSelectedItemChanged: onChanged as void Function(int)?,
               children: animationSpeeds.map((x) => Text('x$x')).toList(),
             ),
           ),
@@ -79,24 +80,26 @@ class _Header extends StatefulWidget {
   _Header({
     this.logoPath,
     this.logoTag,
+    this.logoWidth = 0.75,
     this.title,
     this.titleTag,
     this.height = 250.0,
     this.logoController,
     this.titleController,
-    @required this.loginTheme,
+    required this.loginTheme,
     this.footer,
   });
 
-  final String logoPath;
-  final String logoTag;
-  final String title;
-  final String titleTag;
+  final String? logoPath;
+  final String? logoTag;
+  final double logoWidth;
+  final String? title;
+  final String? titleTag;
   final double height;
   final LoginTheme loginTheme;
-  final AnimationController logoController;
-  final AnimationController titleController;
-  final String footer;
+  final AnimationController? logoController;
+  final AnimationController? titleController;
+  final String? footer;
 
   @override
   __HeaderState createState() => __HeaderState();
@@ -115,7 +118,7 @@ class __HeaderState extends State<_Header> {
     final renderParagraph = RenderParagraph(
       TextSpan(
         text: widget.title,
-        style: theme.textTheme.headline3.copyWith(
+        style: theme.textTheme.headline3!.copyWith(
           fontSize: widget.loginTheme.beforeHeroFontSize,
         ),
       ),
@@ -155,24 +158,25 @@ class __HeaderState extends State<_Header> {
             gap,
         kMaxLogoHeight);
     final displayLogo = widget.logoPath != null && logoHeight >= kMinLogoHeight;
+    final cardWidth = min(MediaQuery.of(context).size.width * 0.75, 360.0);
 
     var logo = displayLogo
         ? Image.asset(
-            widget.logoPath,
+            widget.logoPath!,
             filterQuality: FilterQuality.high,
             height: logoHeight,
-            width: MediaQuery.of(context).size.width * 0.75,
+            width: widget.logoWidth * cardWidth,
           )
         : NullWidget();
 
     if (widget.logoTag != null) {
       logo = Hero(
-        tag: widget.logoTag,
+        tag: widget.logoTag!,
         child: logo,
       );
     }
 
-    Widget title;
+    Widget? title;
     if (widget.titleTag != null && !DartHelper.isNullOrEmpty(widget.title)) {
       title = HeroText(
         widget.title,
@@ -185,7 +189,7 @@ class __HeaderState extends State<_Header> {
       );
     } else if (!DartHelper.isNullOrEmpty(widget.title)) {
       title = Text(
-        widget.title,
+        widget.title!,
         key: kTitleKey,
         style: theme.textTheme.headline3,
       );
@@ -222,25 +226,29 @@ class __HeaderState extends State<_Header> {
 
 class FlutterLogin extends StatefulWidget {
   FlutterLogin(
-      {Key key,
-      @required this.onSignup,
-      @required this.onLogin,
-      @required this.onRecoverPassword,
+      {Key? key,
+      required this.onSignup,
+      required this.onLogin,
+      required this.onRecoverPassword,
       this.title,
       this.logo,
       this.messages,
       this.theme,
-      this.emailValidator,
+      this.userValidator,
       this.passwordValidator,
       this.onSubmitAnimationCompleted,
       this.logoTag,
+      this.userType = LoginUserType.email,
       this.titleTag,
       this.showDebugButtons = false,
       this.loginProviders = const <LoginProvider>[],
       this.hideForgotPasswordButton = false,
       this.hideSignUpButton = false,
       this.loginAfterSignUp = true,
-      this.footer})
+      this.footer,
+      this.hideProvidersTitle = false,
+      this.disableCustomPageTransformer = false,
+      this.navigateBackAfterRecovery = false})
       : super(key: key);
 
   /// Called when the user hit the submit button when in sign up mode
@@ -248,6 +256,10 @@ class FlutterLogin extends StatefulWidget {
 
   /// Called when the user hit the submit button when in login mode
   final AuthCallback onLogin;
+
+  /// [LoginUserType] can be email, name or phone, by default is email. It will change how
+  /// the edit text autofill and behave accordingly to your choice
+  final LoginUserType userType;
 
   /// list of LoginProvider each have an icon and a callback that will be Called when
   /// the user hit the provider icon button
@@ -258,39 +270,39 @@ class FlutterLogin extends StatefulWidget {
   final RecoverCallback onRecoverPassword;
 
   /// The large text above the login [Card], usually the app or company name
-  final String title;
+  final String? title;
 
   /// The path to the asset image that will be passed to the `Image.asset()`
-  final String logo;
+  final String? logo;
 
   /// Describes all of the labels, text hints, button texts and other auth
   /// descriptions
-  final LoginMessages messages;
+  final LoginMessages? messages;
 
   /// FlutterLogin's theme. If not specified, it will use the default theme as
   /// shown in the demo gifs and use the colorsheme in the closest `Theme`
   /// widget
-  final LoginTheme theme;
+  final LoginTheme? theme;
 
   /// Email validating logic, Returns an error string to display if the input is
   /// invalid, or null otherwise
-  final FormFieldValidator<String> emailValidator;
+  final FormFieldValidator<String>? userValidator;
 
-  /// Same as [emailValidator] but for password
-  final FormFieldValidator<String> passwordValidator;
+  /// Same as [userValidator] but for password
+  final FormFieldValidator<String>? passwordValidator;
 
   /// Called after the submit animation's completed. Put your route transition
   /// logic here. Recommend to use with [logoTag] and [titleTag]
-  final Function onSubmitAnimationCompleted;
+  final Function? onSubmitAnimationCompleted;
 
   /// Hero tag for logo image. If not specified, it will simply fade out when
   /// changing route
-  final String logoTag;
+  final String? logoTag;
 
   /// Hero tag for title text. Need to specify `LoginTheme.beforeHeroFontSize`
   /// and `LoginTheme.afterHeroFontSize` if you want different font size before
   /// and after hero animation
-  final String titleTag;
+  final String? titleTag;
 
   /// Display the debug buttons to quickly forward/reverse login animations. In
   /// release mode, this will be overrided to false regardless of the value
@@ -307,17 +319,27 @@ class FlutterLogin extends StatefulWidget {
   final bool loginAfterSignUp;
 
   /// Optional footer text for example a copyright notice
-  final String footer;
+  final String? footer;
+
+  /// Hide the title above the login providers. If no providers are set this is uneffective
+  final bool hideProvidersTitle;
+
+  /// Disable the page transformation between switching authentication modes.
+  /// Fixes #97 if disabled. https://github.com/NearHuscarl/flutter_login/issues/97
+  final bool disableCustomPageTransformer;
+
+  /// Navigate back to the login screen after recovery of password.
+  final bool navigateBackAfterRecovery;
 
   static final FormFieldValidator<String> defaultEmailValidator = (value) {
-    if (value.isEmpty || !Regex.email.hasMatch(value)) {
+    if (value!.isEmpty || !Regex.email.hasMatch(value)) {
       return 'Invalid email!';
     }
     return null;
   };
 
   static final FormFieldValidator<String> defaultPasswordValidator = (value) {
-    if (value.isEmpty || value.length <= 2) {
+    if (value!.isEmpty || value.length <= 2) {
       return 'Password is too short!';
     }
     return null;
@@ -331,9 +353,9 @@ class _FlutterLoginState extends State<FlutterLogin>
     with TickerProviderStateMixin {
   final GlobalKey<AuthCardState> authCardKey = GlobalKey();
   static const loadingDuration = Duration(milliseconds: 400);
-  AnimationController _loadingController;
-  AnimationController _logoController;
-  AnimationController _titleController;
+  AnimationController? _loadingController;
+  AnimationController? _logoController;
+  AnimationController? _titleController;
   double _selectTimeDilation = 1.0;
 
   @override
@@ -345,12 +367,12 @@ class _FlutterLoginState extends State<FlutterLogin>
       duration: loadingDuration,
     )..addStatusListener((status) {
         if (status == AnimationStatus.forward) {
-          _logoController.forward();
-          _titleController.forward();
+          _logoController!.forward();
+          _titleController!.forward();
         }
         if (status == AnimationStatus.reverse) {
-          _logoController.reverse();
-          _titleController.reverse();
+          _logoController!.reverse();
+          _titleController!.reverse();
         }
       });
     _logoController = AnimationController(
@@ -363,24 +385,24 @@ class _FlutterLoginState extends State<FlutterLogin>
     );
 
     Future.delayed(const Duration(seconds: 1), () {
-      _loadingController.forward();
+      _loadingController!.forward();
     });
   }
 
   @override
   void dispose() {
-    _loadingController.dispose();
-    _logoController.dispose();
-    _titleController.dispose();
+    _loadingController!.dispose();
+    _logoController!.dispose();
+    _titleController!.dispose();
     super.dispose();
   }
 
   void _reverseHeaderAnimation() {
     if (widget.logoTag == null) {
-      _logoController.reverse();
+      _logoController!.reverse();
     }
     if (widget.titleTag == null) {
-      _titleController.reverse();
+      _titleController!.reverse();
     }
   }
 
@@ -391,6 +413,7 @@ class _FlutterLoginState extends State<FlutterLogin>
       height: height,
       logoPath: widget.logo,
       logoTag: widget.logoTag,
+      logoWidth: widget.theme?.logoWidth ?? 0.75,
       title: widget.title,
       titleTag: widget.titleTag,
       loginTheme: loginTheme,
@@ -437,19 +460,20 @@ class _FlutterLoginState extends State<FlutterLogin>
           MaterialButton(
             materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
             color: Colors.blue,
-            onPressed: () => authCardKey.currentState.runLoadingAnimation(),
+            onPressed: () => authCardKey.currentState!.runLoadingAnimation(),
             child: Text('LOADING', style: textStyle),
           ),
           MaterialButton(
             materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
             color: Colors.orange,
-            onPressed: () => authCardKey.currentState.runChangePageAnimation(),
+            onPressed: () => authCardKey.currentState!.runChangePageAnimation(),
             child: Text('PAGE', style: textStyle),
           ),
           MaterialButton(
             materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
             color: Colors.red,
-            onPressed: () => authCardKey.currentState.runChangeRouteAnimation(),
+            onPressed: () =>
+                authCardKey.currentState!.runChangeRouteAnimation(),
             child: Text('NAV', style: textStyle),
           ),
         ],
@@ -457,11 +481,17 @@ class _FlutterLoginState extends State<FlutterLogin>
     );
   }
 
-  ThemeData _mergeTheme({ThemeData theme, LoginTheme loginTheme}) {
+  ThemeData _mergeTheme(
+      {required ThemeData theme, required LoginTheme loginTheme}) {
+    final blackOrWhite =
+        theme.brightness == Brightness.light ? Colors.black54 : Colors.white;
+    final primaryOrWhite = theme.brightness == Brightness.light
+        ? theme.primaryColor
+        : Colors.white;
     final originalPrimaryColor = loginTheme.primaryColor ?? theme.primaryColor;
     final primaryDarkShades = getDarkShades(originalPrimaryColor);
     final primaryColor = primaryDarkShades.length == 1
-        ? lighten(primaryDarkShades.first)
+        ? lighten(primaryDarkShades.first!)
         : primaryDarkShades.first;
     final primaryColorDark = primaryDarkShades.length >= 3
         ? primaryDarkShades[2]
@@ -469,25 +499,25 @@ class _FlutterLoginState extends State<FlutterLogin>
     final accentColor = loginTheme.accentColor ?? theme.accentColor;
     final errorColor = loginTheme.errorColor ?? theme.errorColor;
     // the background is a dark gradient, force to use white text if detect default black text color
-    final isDefaultBlackText = theme.textTheme.headline3.color ==
-        Typography.blackMountainView.headline3.color;
-    final titleStyle = theme.textTheme.headline3
+    final isDefaultBlackText = theme.textTheme.headline3!.color ==
+        Typography.blackMountainView.headline3!.color;
+    final titleStyle = theme.textTheme.headline3!
         .copyWith(
           color: loginTheme.accentColor ??
               (isDefaultBlackText
                   ? Colors.white
-                  : theme.textTheme.headline3.color),
+                  : theme.textTheme.headline3!.color),
           fontSize: loginTheme.beforeHeroFontSize,
           fontWeight: FontWeight.w300,
         )
         .merge(loginTheme.titleStyle);
-    final textStyle = theme.textTheme.bodyText2
-        .copyWith(color: Colors.black54)
+    final textStyle = theme.textTheme.bodyText2!
+        .copyWith(color: blackOrWhite)
         .merge(loginTheme.bodyStyle);
-    final textFieldStyle = theme.textTheme.subtitle1
-        .copyWith(color: Colors.black.withOpacity(.65), fontSize: 14)
+    final textFieldStyle = theme.textTheme.subtitle1!
+        .copyWith(color: blackOrWhite, fontSize: 14)
         .merge(loginTheme.textFieldStyle);
-    final buttonStyle = theme.textTheme.button
+    final buttonStyle = theme.textTheme.button!
         .copyWith(color: Colors.white)
         .merge(loginTheme.buttonStyle);
     final cardTheme = loginTheme.cardTheme;
@@ -496,6 +526,14 @@ class _FlutterLoginState extends State<FlutterLogin>
     final roundBorderRadius = BorderRadius.circular(100);
 
     LoginThemeHelper.loginTextStyle = titleStyle;
+
+    var labelStyle;
+
+    if (loginTheme.primaryColorAsInputLabel) {
+      labelStyle = TextStyle(color: primaryColor);
+    } else {
+      labelStyle = TextStyle(color: blackOrWhite);
+    }
 
     return theme.copyWith(
       primaryColor: primaryColor,
@@ -514,13 +552,13 @@ class _FlutterLoginState extends State<FlutterLogin>
         filled: inputTheme.filled,
         fillColor: inputTheme.fillColor ??
             Color.alphaBlend(
-              primaryColor.withOpacity(.07),
+              primaryOrWhite.withOpacity(.07),
               Colors.grey.withOpacity(.04),
             ),
         contentPadding: inputTheme.contentPadding ??
             const EdgeInsets.symmetric(vertical: 4.0),
         errorStyle: inputTheme.errorStyle ?? TextStyle(color: errorColor),
-        labelStyle: inputTheme.labelStyle,
+        labelStyle: inputTheme.labelStyle ?? labelStyle,
         enabledBorder: inputTheme.enabledBorder ??
             inputTheme.border ??
             OutlineInputBorder(
@@ -530,7 +568,7 @@ class _FlutterLoginState extends State<FlutterLogin>
         focusedBorder: inputTheme.focusedBorder ??
             inputTheme.border ??
             OutlineInputBorder(
-              borderSide: BorderSide(color: primaryColor, width: 1.5),
+              borderSide: BorderSide(color: primaryColor!, width: 1.5),
               borderRadius: roundBorderRadius,
             ),
         errorBorder: inputTheme.errorBorder ??
@@ -548,7 +586,7 @@ class _FlutterLoginState extends State<FlutterLogin>
         disabledBorder: inputTheme.disabledBorder ?? inputTheme.border,
       ),
       floatingActionButtonTheme: theme.floatingActionButtonTheme.copyWith(
-        backgroundColor: buttonTheme?.backgroundColor ?? primaryColor,
+        backgroundColor: buttonTheme.backgroundColor ?? primaryColor,
         splashColor: buttonTheme.splashColor ?? theme.accentColor,
         elevation: buttonTheme.elevation ?? 4.0,
         highlightElevation: buttonTheme.highlightElevation ?? 2.0,
@@ -575,17 +613,17 @@ class _FlutterLoginState extends State<FlutterLogin>
     const cardInitialHeight = 300;
     final cardTopPosition = deviceSize.height / 2 - cardInitialHeight / 2;
     final headerHeight = cardTopPosition - headerMargin;
-    final emailValidator =
-        widget.emailValidator ?? FlutterLogin.defaultEmailValidator;
+    final userValidator =
+        widget.userValidator ?? FlutterLogin.defaultEmailValidator;
     final passwordValidator =
         widget.passwordValidator ?? FlutterLogin.defaultPasswordValidator;
 
     Widget footerWidget = SizedBox();
     if (widget.footer != null) {
       footerWidget = Padding(
-        padding: EdgeInsets.only(bottom: 10),
+        padding: EdgeInsets.only(bottom: loginTheme.footerBottomPadding),
         child: Text(
-          widget.footer,
+          widget.footer!,
           style: loginTheme.footerTextStyle,
         ),
       );
@@ -633,9 +671,10 @@ class _FlutterLoginState extends State<FlutterLogin>
                     Positioned(
                       child: AuthCard(
                         key: authCardKey,
+                        userType: widget.userType,
                         padding: EdgeInsets.only(top: cardTopPosition),
                         loadingController: _loadingController,
-                        emailValidator: emailValidator,
+                        userValidator: userValidator,
                         passwordValidator: passwordValidator,
                         onSubmit: _reverseHeaderAnimation,
                         onSubmitCompleted: widget.onSubmitAnimationCompleted,
@@ -643,6 +682,12 @@ class _FlutterLoginState extends State<FlutterLogin>
                         hideForgotPasswordButton:
                             widget.hideForgotPasswordButton,
                         loginAfterSignUp: widget.loginAfterSignUp,
+                        hideProvidersTitle: widget.hideProvidersTitle,
+                        disableCustomPageTransformer:
+                            widget.disableCustomPageTransformer,
+                        loginTheme: widget.theme,
+                        navigateBackAfterRecovery:
+                            widget.navigateBackAfterRecovery,
                       ),
                     ),
                     Positioned(