+++ lib/src/providers/login_messages.dart
@@ -1,25 +1,25 @@
 import 'package:flutter/material.dart';
 
 class LoginMessages with ChangeNotifier {
-  LoginMessages({
-    this.usernameHint = defaultUsernameHint,
-    this.passwordHint = defaultPasswordHint,
-    this.confirmPasswordHint = defaultConfirmPasswordHint,
-    this.forgotPasswordButton = defaultForgotPasswordButton,
-    this.loginButton = defaultLoginButton,
-    this.signupButton = defaultSignupButton,
-    this.recoverPasswordButton = defaultRecoverPasswordButton,
-    this.recoverPasswordIntro = defaultRecoverPasswordIntro,
-    this.recoverPasswordDescription = defaultRecoverPasswordDescription,
-    this.goBackButton = defaultGoBackButton,
-    this.confirmPasswordError = defaultConfirmPasswordError,
-    this.recoverPasswordSuccess = defaultRecoverPasswordSuccess,
-    this.flushbarTitleError = defaultflushbarTitleError,
-    this.flushbarTitleSuccess = defaultflushbarTitleSuccess,
-    this.signUpSuccess = defaultSignUpSuccess,
-  });
+  LoginMessages(
+      {this.userHint = defaultUserHint,
+      this.passwordHint = defaultPasswordHint,
+      this.confirmPasswordHint = defaultConfirmPasswordHint,
+      this.forgotPasswordButton = defaultForgotPasswordButton,
+      this.loginButton = defaultLoginButton,
+      this.signupButton = defaultSignupButton,
+      this.recoverPasswordButton = defaultRecoverPasswordButton,
+      this.recoverPasswordIntro = defaultRecoverPasswordIntro,
+      this.recoverPasswordDescription = defaultRecoverPasswordDescription,
+      this.goBackButton = defaultGoBackButton,
+      this.confirmPasswordError = defaultConfirmPasswordError,
+      this.recoverPasswordSuccess = defaultRecoverPasswordSuccess,
+      this.flushbarTitleError = defaultflushbarTitleError,
+      this.flushbarTitleSuccess = defaultflushbarTitleSuccess,
+      this.signUpSuccess = defaultSignUpSuccess,
+      this.providersTitle = defaultProvidersTitle});
 
-  static const defaultUsernameHint = 'Email';
+  static const defaultUserHint = 'Email';
   static const defaultPasswordHint = 'Password';
   static const defaultConfirmPasswordHint = 'Confirm Password';
   static const defaultForgotPasswordButton = 'Forgot Password?';
@@ -35,9 +35,11 @@ class LoginMessages with ChangeNotifier {
   static const defaultflushbarTitleSuccess = 'Success';
   static const defaultflushbarTitleError = 'Error';
   static const defaultSignUpSuccess = 'An activation link has been sent';
+  static const defaultProvidersTitle = 'or login with';
 
-  /// Hint text of the user name [TextField]
-  final String usernameHint;
+  /// Hint text of the userHint [TextField]
+  /// By default is Email
+  final String userHint;
 
   /// Hint text of the password [TextField]
   final String passwordHint;
@@ -82,4 +84,7 @@ class LoginMessages with ChangeNotifier {
 
   /// The success message to show after signing up
   final String signUpSuccess;
+
+  /// The string shown above the Providers buttons
+  final String providersTitle;
 }