+++ test/flutter_login_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter_login/src/constants.dart';
 import 'package:flutter_login/src/widgets/animated_text.dart';
 
 void main() {
-  final TestWidgetsFlutterBinding binding =
-      TestWidgetsFlutterBinding.ensureInitialized();
+  final binding = TestWidgetsFlutterBinding.ensureInitialized()
+      as TestWidgetsFlutterBinding;
 
   void setScreenSize(Size size) {
     binding.window.physicalSizeTestValue = size;
@@ -34,7 +34,7 @@ void main() {
     await tester.pumpAndSettle();
 
     // TODO: put error messages into variables
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid email!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid email!');
 
     // missing '@'
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -43,7 +43,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid email!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid email!');
 
     // missing the part before '@'
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -52,7 +52,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid email!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid email!');
 
     // missing the part after '@'
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -61,7 +61,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid email!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid email!');
 
     // missing domain extension (.com, .org...)
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -70,7 +70,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid email!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid email!');
 
     // valid email based on default validator
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -79,7 +79,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, null);
+    expect(nameTextFieldWidget(tester).decoration!.errorText, null);
   });
 
   testWidgets(
@@ -97,7 +97,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText,
+    expect(passwordTextFieldWidget(tester).decoration!.errorText,
         'Password is too short!');
 
     // too short
@@ -107,7 +107,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText,
+    expect(passwordTextFieldWidget(tester).decoration!.errorText,
         'Password is too short!');
 
     // valid password based on default validator
@@ -117,7 +117,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, null);
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, null);
 
     // valid password based on default validator
     await simulateOpenSoftKeyboard(tester, defaultFlutterLogin());
@@ -127,7 +127,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, null);
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, null);
   });
 
   testWidgets('Confirm password field throws error if not match with password',
@@ -151,8 +151,8 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, null);
-    expect(confirmPasswordTextFieldWidget(tester).decoration.errorText,
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, null);
+    expect(confirmPasswordTextFieldWidget(tester).decoration!.errorText,
         LoginMessages.defaultConfirmPasswordError);
 
     // match
@@ -164,17 +164,17 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, null);
-    expect(confirmPasswordTextFieldWidget(tester).decoration.errorText, null);
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, null);
+    expect(confirmPasswordTextFieldWidget(tester).decoration!.errorText, null);
   });
 
-  testWidgets('Custom emailValidator should show error when return a string',
+  testWidgets('Custom userValidator should show error when return a string',
       (WidgetTester tester) async {
     final loginBuilder = () => widget(FlutterLogin(
           onSignup: (data) => null,
           onLogin: (data) => null,
           onRecoverPassword: (data) => null,
-          emailValidator: (value) => value.endsWith('.com') ? null : 'Invalid!',
+          userValidator: (value) => value!.endsWith('.com') ? null : 'Invalid!',
         ));
     await tester.pumpWidget(loginBuilder());
     await tester.pumpAndSettle(loadingAnimationDuration);
@@ -186,7 +186,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, 'Invalid!');
+    expect(nameTextFieldWidget(tester).decoration!.errorText, 'Invalid!');
 
     // valid value
     await simulateOpenSoftKeyboard(tester, loginBuilder());
@@ -195,7 +195,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).decoration.errorText, null);
+    expect(nameTextFieldWidget(tester).decoration!.errorText, null);
   });
 
   testWidgets('Custom passwordValidator should show error when return a string',
@@ -204,7 +204,7 @@ void main() {
           onSignup: (data) => null,
           onLogin: (data) => null,
           onRecoverPassword: (data) => null,
-          passwordValidator: (value) => value.length == 5 ? null : 'Invalid!',
+          passwordValidator: (value) => value!.length == 5 ? null : 'Invalid!',
         ));
     await tester.pumpWidget(loginBuilder());
     await tester.pumpAndSettle(loadingAnimationDuration);
@@ -216,7 +216,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, 'Invalid!');
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, 'Invalid!');
 
     // valid value
     await simulateOpenSoftKeyboard(tester, loginBuilder());
@@ -225,7 +225,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(passwordTextFieldWidget(tester).decoration.errorText, null);
+    expect(passwordTextFieldWidget(tester).decoration!.errorText, null);
   });
 
   testWidgets('Password recovery should show success message if email is valid',
@@ -236,7 +236,7 @@ void main() {
           onLogin: (data) => null,
           onRecoverPassword: (data) =>
               users.contains(data) ? null : Future.value('User not exists'),
-          emailValidator: (value) => null,
+          userValidator: (value) => null,
         ));
     await tester.pumpWidget(loginBuilder());
     await tester.pumpAndSettle(loadingAnimationDuration);
@@ -281,7 +281,7 @@ void main() {
           onLogin: (data) => null,
           onRecoverPassword: (data) => null,
           messages: LoginMessages(
-            usernameHint: 'Username',
+            userHint: 'User',
             passwordHint: 'Pass',
             confirmPasswordHint: 'Confirm',
             loginButton: 'LOG IN',
@@ -299,15 +299,15 @@ void main() {
     await tester.pumpAndSettle(loadingAnimationDuration);
 
     var nameTextField = nameTextFieldWidget(tester);
-    expect(nameTextField.decoration.labelText, 'Username');
-    expect(find.text('Username'), findsOneWidget);
+    expect(nameTextField.decoration!.labelText, 'User');
+    expect(find.text('User'), findsOneWidget);
 
     final passwordTextField = passwordTextFieldWidget(tester);
-    expect(passwordTextField.decoration.labelText, 'Pass');
+    expect(passwordTextField.decoration!.labelText, 'Pass');
     expect(find.text('Pass'), findsOneWidget);
 
     final confirmPasswordTextField = confirmPasswordTextFieldWidget(tester);
-    expect(confirmPasswordTextField.decoration.labelText, 'Confirm');
+    expect(confirmPasswordTextField.decoration!.labelText, 'Confirm');
     expect(find.text('Confirm'), findsOneWidget);
 
     var submitButton = submitButtonWidget();
@@ -335,7 +335,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    expect(confirmPasswordTextFieldWidget(tester).decoration.errorText,
+    expect(confirmPasswordTextFieldWidget(tester).decoration!.errorText,
         'Not match!');
 
     // Go to forgot password page
@@ -343,8 +343,8 @@ void main() {
     await tester.pumpAndSettle();
 
     nameTextField = nameTextFieldWidget(tester);
-    expect(nameTextField.decoration.labelText, 'Username');
-    expect(find.text('Username'), findsOneWidget);
+    expect(nameTextField.decoration!.labelText, 'User');
+    expect(find.text('User'), findsOneWidget);
 
     submitButton = submitButtonWidget();
     expect(submitButton.text, 'HELP ME');
@@ -474,7 +474,7 @@ void main() {
           onSignup: (data) => null,
           onLogin: mockCallback.onLogin,
           onRecoverPassword: (data) => null,
-          emailValidator: mockCallback.emailValidator,
+          userValidator: mockCallback.userValidator,
           passwordValidator: mockCallback.passwordValidator,
           onSubmitAnimationCompleted: mockCallback.onSubmitAnimationCompleted,
         ));
@@ -495,7 +495,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator('invalid-name'),
+      mockCallback.userValidator('invalid-name'),
       mockCallback.passwordValidator(user.password),
     ]);
     verifyNever(mockCallback.onLogin(any));
@@ -513,7 +513,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator(invalidUser.name),
+      mockCallback.userValidator(invalidUser.name),
       mockCallback.passwordValidator(invalidUser.password),
       mockCallback.onLogin(any),
     ]);
@@ -531,7 +531,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator(user.name),
+      mockCallback.userValidator(user.name),
       mockCallback.passwordValidator(user.password),
       mockCallback.onLogin(any),
       mockCallback.onSubmitAnimationCompleted(),
@@ -547,7 +547,7 @@ void main() {
           onLogin: (data) => null,
           onSignup: mockCallback.onSignup,
           onRecoverPassword: (data) => null,
-          emailValidator: mockCallback.emailValidator,
+          userValidator: mockCallback.userValidator,
           passwordValidator: mockCallback.passwordValidator,
           onSubmitAnimationCompleted: mockCallback.onSubmitAnimationCompleted,
         ));
@@ -573,7 +573,7 @@ void main() {
     clickSubmitButton();
     await tester.pumpAndSettle();
 
-    verifyNever(mockCallback.emailValidator(invalidUser.name));
+    verifyNever(mockCallback.userValidator(invalidUser.name));
     verifyNever(mockCallback.passwordValidator(invalidUser.password));
     verifyNever(mockCallback.onSignup(any));
     verifyNever(mockCallback.onSubmitAnimationCompleted());
@@ -592,7 +592,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator('invalid-name'),
+      mockCallback.userValidator('invalid-name'),
       mockCallback.passwordValidator(user.password),
     ]);
     verifyNever(mockCallback.onSignup(any));
@@ -613,7 +613,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator(invalidUser.name),
+      mockCallback.userValidator(invalidUser.name),
       mockCallback.passwordValidator(invalidUser.password),
       mockCallback.onSignup(any),
     ]);
@@ -633,7 +633,7 @@ void main() {
     await tester.pumpAndSettle();
 
     verifyInOrder([
-      mockCallback.emailValidator(user.name),
+      mockCallback.userValidator(user.name),
       mockCallback.passwordValidator(user.password),
       mockCallback.onSignup(any),
       mockCallback.onSubmitAnimationCompleted(),
@@ -663,14 +663,14 @@ void main() {
     clickForgotPasswordButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).controller.text, 'near@gmail.com');
+    expect(nameTextFieldWidget(tester).controller!.text, 'near@gmail.com');
 
     clickGoBackButton();
     await tester.pumpAndSettle();
 
-    expect(nameTextFieldWidget(tester).controller.text, 'near@gmail.com');
-    expect(passwordTextFieldWidget(tester).controller.text, '12345');
-    expect(confirmPasswordTextFieldWidget(tester).controller.text, 'abcde');
+    expect(nameTextFieldWidget(tester).controller!.text, 'near@gmail.com');
+    expect(passwordTextFieldWidget(tester).controller!.text, '12345');
+    expect(confirmPasswordTextFieldWidget(tester).controller!.text, 'abcde');
   });
 
   // https://github.com/NearHuscarl/flutter_login/issues/20
@@ -732,7 +732,7 @@ void main() {
           onSignup: (data) => null,
           onLogin: (data) => null,
           onRecoverPassword: (data) => null,
-          passwordValidator: (value) => value.length == 5 ? null : 'Invalid!',
+          passwordValidator: (value) => value!.length == 5 ? null : 'Invalid!',
           hideSignUpButton: true,
           hideForgotPasswordButton: true,
           messages: LoginMessages(
@@ -744,54 +744,6 @@ void main() {
     await tester.pumpAndSettle(loadingAnimationDuration);
     expect(find.text('REGISTER'), findsNothing);
     expect(find.text('Forgot huh?'), findsNothing);
-  });
-
-  testWidgets('LoginProvider buttons should login or throw error.',
-      (WidgetTester tester) async {
-    final loginBuilder = () => widget(FlutterLogin(
-          onSignup: (data) => null,
-          onLogin: (data) => null,
-          onRecoverPassword: (data) => null,
-          passwordValidator: (value) => value.length == 5 ? null : 'Invalid!',
-          loginProviders: [
-            LoginProvider(
-              icon: Icons.check,
-              callback: () async {
-                return 'No problems!';
-              },
-            ),
-            LoginProvider(
-              icon: Icons.warning,
-              callback: () async {
-                return 'Failed to login';
-              },
-            ),
-          ],
-        ));
-    await tester.pumpWidget(loginBuilder());
-    await tester.pumpAndSettle(loadingAnimationDuration);
-    expect(find.byIcon(Icons.check), findsOneWidget);
-    expect(find.byIcon(Icons.warning), findsOneWidget);
-
-    await tester.tap(find.byIcon(Icons.warning));
-
-    // Because of multiple animations, in order to get to the flushbar we need
-    // to pump the animations two times.
-    await tester.pump();
-    await tester.pump(const Duration(seconds: 4));
-    await tester.pump(const Duration(seconds: 4));
-    expect(find.text('Failed to login'), findsOneWidget);
-    await tester.pumpAndSettle();
-
-    await tester.tap(find.byIcon(Icons.check));
-
-    // Because of multiple animations, in order to get to the flushbar we need
-    // to pump the animations two times.
-    await tester.pump();
-    await tester.pump(const Duration(seconds: 4));
-    await tester.pump(const Duration(seconds: 4));
-    expect(find.text('No problems!'), findsOneWidget);
-    await tester.pumpAndSettle();
   });
 
   testWidgets(